// ui-v2/packages/libs/4-apps/src/statement/models/statement-model.ts
export interface StatementModel {
  avg_affected_rows?: number
  avg_backoff_time?: number
  avg_commit_backoff_time?: number
  avg_commit_time?: number
  avg_compile_latency?: number
  avg_cop_process_time?: number
  avg_cop_wait_time?: number
  avg_disk?: number
  avg_get_commit_ts_time?: number
  avg_latency?: number
  avg_local_latch_wait_time?: number
  avg_mem?: number
  avg_parse_latency?: number
  avg_prewrite_regions?: number
  avg_prewrite_time?: number
  avg_process_time?: number
  avg_processed_keys?: number
  avg_resolve_lock_time?: number
  avg_rocksdb_block_cache_hit_count?: number
  avg_rocksdb_block_read_byte?: number
  avg_rocksdb_block_read_count?: number
  avg_rocksdb_delete_skipped_count?: number
  avg_rocksdb_key_skipped_count?: number
  avg_ru?: number
  avg_time_queued_by_rc?: number
  avg_total_keys?: number
  avg_txn_retry?: number
  avg_wait_time?: number
  avg_write_keys?: number
  avg_write_size?: number
  binary_plan?: string
  binary_plan_json?: string
  binary_plan_text?: string
  digest?: string
  digest_text?: string
  exec_count?: number
  first_seen?: number
  index_names?: string
  last_seen?: number
  max_backoff_time?: number
  max_commit_backoff_time?: number
  max_commit_time?: number
  max_compile_latency?: number
  max_cop_process_time?: number
  max_cop_wait_time?: number
  max_disk?: number
  max_get_commit_ts_time?: number
  max_latency?: number
  max_local_latch_wait_time?: number
  max_mem?: number
  max_parse_latency?: number
  max_prewrite_regions?: number
  max_prewrite_time?: number
  max_process_time?: number
  max_processed_keys?: number
  max_resolve_lock_time?: number
  max_rocksdb_block_cache_hit_count?: number
  max_rocksdb_block_read_byte?: number
  max_rocksdb_block_read_count?: number
  max_rocksdb_delete_skipped_count?: number
  max_rocksdb_key_skipped_count?: number
  max_ru?: number
  max_time_queued_by_rc?: number
  max_total_keys?: number
  max_txn_retry?: number
  max_wait_time?: number
  max_write_keys?: number
  max_write_size?: number
  min_latency?: number
  plan?: string
  plan_cache_hits?: number
  plan_can_be_bound?: boolean
  plan_count?: number
  plan_digest?: string
  plan_hint?: string
  prev_sample_text?: string
  query_sample_text?: string
  related_schemas?: string
  resource_group?: string
  sample_user?: string
  schema_name?: string
  stmt_type?: string
  sum_backoff_times?: number
  sum_cop_task_num?: number
  sum_errors?: number
  sum_latency?: number
  sum_ru?: number
  sum_warnings?: number
  summary_begin_time?: number
  summary_end_time?: number
  table_names?: string
}